+++ lib/src/treestructureddata/treestructureddata.dart
@@ -24,11 +24,16 @@ class TreeStructuredData extends TreeNode<Name, Value> {
       [Value priority])
       : this._(null, children, priority);
 
-  factory TreeStructuredData.fromJson(json,
-      [priority, Map<ServerValue, Value> serverValues]) {
+  factory TreeStructuredData.fromJson(json, [priority]) {
     if (json == null) {
       return new TreeStructuredData();
     }
+    if (json is !Map&&json is !bool&&json is !num&&json is !String&&json is !List) {
+      try {
+        json = json.toJson();
+      } on NoSuchMethodError {}
+    }
+
     if (json is Map && json.containsKey(".priority")) {
       priority = json[".priority"];
     }
@@ -38,17 +43,17 @@ class TreeStructuredData extends TreeNode<Name, Value> {
       json = json[".value"];
     }
 
+    if (json is List) json = json.asMap();
     if (json is! Map || json.containsKey(".sv")) {
       var value = new Value(json);
-      if (value is ServerValue) value = serverValues[value];
       return new TreeStructuredData.leaf(value, priority);
     }
 
     var children = new Map<Name, TreeStructuredData>.fromIterable(
-        json.keys.where((k) => !k.startsWith(".")),
+        json.keys.map((k)=>k.toString()).where((k) => !k.startsWith(".")),
         key: (k) => new Name(k),
         value: (k) =>
-            new TreeStructuredData.fromJson(json[k], null, serverValues));
+            new TreeStructuredData.fromJson(json[k], null));
 
     return new TreeStructuredData.nonLeaf(children, priority);
   }