+++ lib/src/treestructureddata/value.dart
@@ -7,11 +7,24 @@ class ServerValue {
   final String type;
   const ServerValue._(this.type);
 
-  static const ServerValue timestamp = const ServerValue._("TIMESTAMP");
+  static const ServerValue timestamp = const ServerValue._("timestamp");
 
-  static const Map<String, ServerValue> values = const {"TIMESTAMP": timestamp};
+  static const Map<String, ServerValue> values = const {"timestamp": timestamp};
 
   Map<String, String> toJson() => {".sv": type};
+
+  static TreeStructuredData resolve(TreeStructuredData value,
+      Map<ServerValue, Value> serverValues) {
+    if (value.isLeaf) {
+      return value.value.value is ServerValue ?
+      new TreeStructuredData.leaf(serverValues[value.value.value]) : value;
+    }
+    var newValue = value.clone();
+    for (var k in newValue.children.keys.toList()) {
+      newValue.children[k] = resolve(newValue.children[k], serverValues);
+    }
+    return newValue;
+  }
 }
 
 class Value implements Comparable<Value> {